feat: Refactor package structure for utils and controllers.
Moved `FilterUtils` to a higher-level `utils` package and relocated `AdminCacheController` and its implementation to a dedicated `cache` package. Adjusted import paths across affected files to reflect these changes. This enhances code organization and maintainability.
