+++ lsadf_admin/src/test/java/com/lsadf/admin/application/unit/utils/FilterUtilsTests.java
@@ -17,7 +17,7 @@ package com.lsadf.admin.application.unit.utils;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.lsadf.admin.application.utils.FilterUtils;
+import com.lsadf.admin.utils.FilterUtils;
 import com.lsadf.core.domain.user.User;
 import com.lsadf.core.infra.utils.DateUtils;
 import com.vaadin.hilla.crud.filter.AndFilter;
